requirements.txt
# ===== CORE =====
python-dateutil>=2.8.2

# ===== NUMPY (BẮT BUỘC < 2.0 vì scikit-surprise) =====
numpy==1.26.4

# ===== DATA =====
pandas>=2.0.3
scipy>=1.10.1

# ===== MACHINE LEARNING =====
scikit-learn>=1.3.0
scikit-surprise==1.1.4

# ===== VISUALIZATION =====
matplotlib>=3.7.2
seaborn>=0.12.2

# ===== WEB UI =====
streamlit>=1.29.0